requirements.txt
# Core AI Frameworks
crewai==1.9.3
crewai-tools==1.9.3
google-genai==1.49.0
google-generativeai

# API and Server (Crucial for Docker)
fastapi==0.129.2
uvicorn==0.41.0
python-dotenv==1.1.1
python-multipart==0.0.22

# Utilities
pydantic==2.11.10
pydantic-settings==2.10.1
requests==2.32.5
pyyaml==6.0.3
databricks-sql-connector>=4.2.5